Renamed provider types, fixed crd object creation for DNSEntry, added crds.yaml
