+++ examples/crds.yaml
@@ -1,0 +1,99 @@
+apiVersion: apiextensions.k8s.io/v1beta1
+kind: CustomResourceDefinition
+metadata:
+  name: dnsentries.dns.gardener.cloud
+spec:
+  additionalPrinterColumns:
+  - JSONPath: .spec.dnsName
+    description: DNS ObjectName
+    name: DNS
+    type: string
+  - JSONPath: .status.state
+    description: Status of DNS entry in cloud provider
+    name: STATUS
+    type: string
+  - JSONPath: .metadata.creationTimestamp
+    name: AGE
+    type: date
+  conversion:
+    strategy: None
+  group: dns.gardener.cloud
+  names:
+    kind: DNSEntry
+    listKind: DNSEntryList
+    plural: dnsentries
+    shortNames:
+    - dnse
+    singular: dnsentry
+  scope: Namespaced
+  subresources:
+    status: {}
+  version: v1alpha1
+  versions:
+  - name: v1alpha1
+    served: true
+    storage: true
+---
+apiVersion: apiextensions.k8s.io/v1beta1
+kind: CustomResourceDefinition
+metadata:
+  name: dnsowners.dns.gardener.cloud
+spec:
+  additionalPrinterColumns:
+  - JSONPath: .spec.ownerId
+    description: Owner Id
+    name: OWNERID
+    type: string
+  - JSONPath: .metadata.creationTimestamp
+    name: AGE
+    type: date
+  conversion:
+    strategy: None
+  group: dns.gardener.cloud
+  names:
+    kind: DNSOwner
+    listKind: DNSOwnerList
+    plural: dnsowners
+    shortNames:
+    - dnso
+    singular: dnsowner
+  scope: Cluster
+  version: v1alpha1
+  versions:
+  - name: v1alpha1
+    served: true
+    storage: true
+---
+apiVersion: apiextensions.k8s.io/v1beta1
+kind: CustomResourceDefinition
+metadata:
+  name: dnsproviders.dns.gardener.cloud
+spec:
+  additionalPrinterColumns:
+  - JSONPath: .status.providerType
+    description: Provider type
+    name: TYPE
+    type: string
+  - JSONPath: .status.state
+    description: Status of DNS provider
+    name: STATUS
+    type: string
+  - JSONPath: .metadata.creationTimestamp
+    name: AGE
+    type: date
+  conversion:
+    strategy: None
+  group: dns.gardener.cloud
+  names:
+    kind: DNSProvider
+    listKind: DNSProviderList
+    plural: dnsproviders
+    shortNames:
+    - dnspr
+    singular: dnsprovider
+  scope: Namespaced
+  version: v1alpha1
+  versions:
+  - name: v1alpha1
+    served: true
+    storage: true
