+++ examples/secret_alicloud_credentials.yaml
@@ -1,0 +1,10 @@
+apiVersion: v1
+kind: Secret
+metadata:
+  name: alicloud-credentials
+  namespace: default
+type: Opaque
+data:
+  # Replace '...' with values encoded as base64.
+  ACCESS_KEY_ID: ...
+  SECRET_ACCESS_KEY: ...
