+++ pkg/controller/provider/aws/execution.go
@@ -14,7 +14,7 @@
  *
  */
 
-package route53
+package aws
 
 import (
 	"github.com/aws/aws-sdk-go/aws"
