+++ pkg/controller/provider/google/controller.go
@@ -14,17 +14,16 @@
  *
  */
 
-package route53
+package google
 
 import (
 	"github.com/gardener/external-dns-management/pkg/dns/provider"
 )
 
-const TYPE_CODE = "AWS"
-const CONTROLLER_NAME = "route53-dns-controller"
+const TYPE_CODE = "google-clouddns"
 
 func init() {
-	provider.DNSController(CONTROLLER_NAME, provider.NewDNSHandlerFactory(TYPE_CODE, NewHandler)).
+	provider.DNSController(TYPE_CODE, provider.NewDNSHandlerFactory(TYPE_CODE, NewHandler)).
 		FinalizerDomain("dns.gardener.cloud").
 		MustRegister(provider.CONTROLLER_GROUP_DNS_CONTROLLERS)
 }
